pyproject.toml
[tool.poetry]
name = "hbuild"
version = "0.1.0"
description = "Hades package build system"
authors = ["Yavuz Rao <yavuz.rao@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
rustworkx = "^0.15.1"
jsonschema = "^4.23.0"
pyyaml = "^6.0.2"
toml = "^0.10.2"
rich = "^13.9.4"
gitpython = "^3.1.43"
pydot = "^3.0.3"
pillow = "^11.0.0"
jsonpickle = "^4.0.1"
argparse-dataclass = "^2.0.0"
dearpygui = "^2.0.0"
docker = "^7.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
